형식 : TTTFunc에서 분할 컴파일을 위한 코드 수정, 가독성을 위한 주석추가
변경 전
 - TTTFunc내에 랭킹관여 함수가 존재하였다.
 - 랭킹관여 함수 주석이 미흡하였다.

변경 후
 - 랭킹 관여 함수를 Ranking.c 파일로 분할하였다.
 - 랭킹 관여 함수를 읽기 쉽게 주석을 추가 하였다.

 *추가적인 사항*
 - 최신화를 위한 fetch시 #31의 문제가 수정된것 같으니 재확인 할 것
 - #27 중 Rank.c 가 아닌 Ranking.c로 수정함
